fix(ci): watch what determines a suite's result, and stop overstating the run
Codex round 9 on #1796. Five findings, all reproduced first.

1. A suite's transitive local imports were unwatched (P1). An `alpha` suite
   rewrote a HELPER that a later `beta` suite imports — not a suite file, so no
   digest covered it — and `beta`, which fails against the committed helper,
   reported ok at exit 0.

   The boundary was wrong, not the list. The rule is now "anything whose content
   can change a suite's result is an input": the watch set folds in each suite's
   transitive first-party import closure, and the exemption importer's closure
   too. Only relative specifiers are followed (a bare one is a node: builtin or a
   lockfile-pinned package); a relative import resolving outside the repository
   fails the gate closed, since it cannot be pinned by a repo-relative digest.
   `gateInputs` now requires the root and throws without it, because defaulting
   would silently shrink the derived set back to the pre-closure one.

   The closure needed a line-anchored scanner: `staticImportSpecifiers` counted
   `import` statements appearing inside STRING LITERALS, so a test file that
   embeds fixture source was credited with dependencies it never had.

2. `sentry-suite-gate-fixtures.mjs` was unrouted (P1) — created by the round-8
   split, matching no arm, so a change to it scheduled Trunk and lint:scripts but
   NEITHER gate suite, despite owning fixture environment isolation, the
   step-summary redirection and the shared harness. The arm now matches
   `scripts/sentry-suite-gate*.mjs`, covering every current and future gate
   module; the routing assertion covers both files the split created.

3. The exemption route accepted any script, not the named one (P2). With
   `tf:test` no-op'd and an unused `decoy` running the importer, a provider suite
   that throws on load was skipped entirely. The `via` alias is now resolved and
   validated exactly, since production-infra-contract runs only `pnpm tf:test`.

4. The summary made a false attestation (P2). It claimed all N entries were
   "reconciled and asserted from their own output" while the exempt entry was
   never spawned. It now counts run and exempt separately and states both, in
   wording true of a normal run without the reader knowing what "exempt" means.

5. Runbook (P2): the direct integrity-suite command joins its sibling in the
   canonical reproduction list.

Two file-size splits, both forced and both proven neutral by diffing the sorted
case-name lists: scripts/sentry-suite-gate-integrity.mjs (the integrity layer,
same seam as the test split) and scripts/check-sentry-suites-in-ci-coverage.mjs
(#1803's prescribed split of the 999-line entry point, 27 cases = 19 + 8). Both
registered in the file-size pin. Entry point 1000 -> 724, gate 1065 -> 865.

Refs #1779, #1803.
